% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencode}
\alias{autoencode}
\title{R interface for the autoencode python function}
\usage{
autoencode(x, python.module, main, test.x = NULL, pretrain_file = "",
  nonmissing_indicator = 1, n_human = 21183L, n_mouse = 21122L,
  shared_size = 15494L, model.species = NULL, out_dir = ".",
  batch_size = 32L, ...)
}
\arguments{
\item{x}{Target data matrix of gene by cell. When pretraining is used, the genes should be the same as the nodes used in the pretrained model. If a node gene is missing is the target dataset, set all values of that gene as 0 in \code{x} and indicate that using \code{nonmissing_indicator}}

\item{python.module}{The python module for the Python package \code{sctransfer}}

\item{main}{A Python main module}

\item{test.x}{Data matrix to evaluate the test error}

\item{pretrain_file}{The pretrained weights file ended with '.hdf5'}

\item{nonmissing_indicator}{A single value 1 or a vector of 0 and 1s to indicate which nodes are missing in the target dataset. Set to 1 for no pretraining.}

\item{model.species}{Should be either 'Human' or 'Mouse' when pretraining is used}

\item{...}{Extra parameters passed to Python module \code{sctransfer} function \code{api} (if no pretraining) or function \code{api_pretrain} (with pretraining).}
}
\value{
A data matrix for the Python autoencoder result
}
\description{
R interface for the autoencode python function
}
