% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeShrinkage.R
\name{computeShrinkage}
\alias{computeShrinkage}
\title{Wrapper function for the empirical Bayes shrinkage step}
\usage{
computeShrinkage(text.file.name, ncores = 1)
}
\arguments{
\item{text.file.name}{Can be either .txt, .csv or .rds files that store the data matrix gene by cell. The rds file can either store the data as a regular matrix or a sparse matrix}

\item{ncores}{number of cores that can be used for the SAVER shrinkage}
}
\value{
RDS saved for the final denoised results
}
\description{
Wrapper function for the empirical Bayes shrinkage step
}
