% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{mapHuman}
\alias{mapHuman}
\title{map given human gene names to our internal unique 64123 human gene IDs}
\usage{
mapHuman(gene.names)
}
\arguments{
\item{gene.names}{a vector of given gene names that need to be mapped to}
}
\description{
map given human gene names to our internal unique 64123 human gene IDs
}
