% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_funs.R
\name{mapMouse}
\alias{mapMouse}
\title{map given mouse gene names to our internal unique 67618 mouse gene IDs}
\usage{
mapMouse(gene.names)
}
\arguments{
\item{gene.names}{a vector of given gene names that need to be mapped to}
}
\description{
map given mouse gene names to our internal unique 67618 mouse gene IDs
}
