% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverx.R
\name{saverx}
\alias{saverx}
\title{the SAVERX function}
\usage{
saverx(text.file.name, data.species = c("Human", "Mouse", "Others"),
  use.pretrain = F, pretrained.weights.file = "",
  model.species = c("Human", "Mouse", "Joint"), model.nodes.ID = NULL,
  ncores = 1, verbose = F, ...)
}
\arguments{
\item{text.file.name}{Can be either .txt, .csv or .rds files that store the data matrix gene by cell. The rds file can either store the data as a regular matrix or a sparse matrix}

\item{data.species}{The species of the dataset}

\item{use.pretrain}{Use a pretrained model or not}

\item{pretrained.weights.file}{If a pretrained model is used, provide the file storing the autoencoder model weights. It should have an extension of ".hdf5" and is the saved weights from the Python package \code{transferdca}}

\item{model.species}{The species of the pretrained model}

\item{model.nodes.ID}{The vector of node IDs of the pretrained model (only needed for the species of the data when the pre-trained model is joint). Set to NULL if running SAVER-X without pretraining.}

\item{ncores}{number of cores that can be used for the SAVER shrinkage}

\item{verbose}{Whether to show more autoencoder optimization progress or not}

\item{...}{more arguments passed to \code{autoFilterCV}}
}
\value{
the final denoised RDS data file saved in the same directory as the input data
}
\description{
Output and intemediate files are stored at the same directory as the input data file \code{text.file.name}. To avoid potential file name conflicts, please make the folder only contain \code{text.file.name}. DO NOT run two SAVER-X tasks for the data file in the same folder.
}
