% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeShrinkage.R
\name{computeShrinkage}
\alias{computeShrinkage}
\title{Wrapper function for the empirical Bayes shrinkage step}
\usage{
computeShrinkage(out.dir, ncores = 1)
}
\arguments{
\item{out.dir}{a directory store all intemediate files. Can be generated automatically by this function is called by \code{computePrediction}}

\item{ncores}{number of cores that can be used for the SAVER shrinkage}
}
\value{
RDS saved for the final denoised results
}
\description{
Wrapper function for the empirical Bayes shrinkage step
}
