% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data_fun.R
\name{preprocessDat}
\alias{preprocessDat}
\title{Pre-process the data for SAVER-X}
\usage{
preprocessDat(text.file.name, data.species = c("Human", "Mouse",
  "Others"), model.species = c("Human", "Mouse", "Joint"),
  model.nodes.ID = NULL)
}
\arguments{
\item{text.file.name}{Can be either .txt, .csv or .rds files that store the data matrix gene by cell. The rds file can either store the data as a regular matrix or a sparse matrix}

\item{data.species}{The species of the dataset}

\item{model.species}{The species of the pretrained model}

\item{model.nodes.ID}{The vector of node IDs of the pretrained model (only needed for the species of the data when the pre-trained model is joint). Set to NULL if running SAVER-X without pretraining.}
}
\value{
Do not return anything but generate a series of intermediate files in the same folder as \code{text.file.name}
}
\description{
Output and intemediate files are stored at the same directory as text.file.name. To avoid potential file name conflicts, please make the folder only contain text.file.name. DO NOT run two SAVER-X tasks for the data file in the same folder.
}
